requirements.txt
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.26
openai>=1.11.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
streamlit>=1.35.0